+++ RiskOfTheAncients2/Equipment/ArmletOfMordiggian.cs
@@ -1,5 +1,7 @@
 ﻿using BepInEx.Configuration;
+using MonoMod.RuntimeDetour;
 using RoR2;
+using RoR2.UI;
 using ROTA2.Buffs;
 using System;
 using UnityEngine;
@@ -18,11 +20,20 @@ namespace ROTA2.Equipment
         public override bool EquipmentIsLunar => true;
         public override bool EquipmentCanBeRandomlyTriggered => false;
         public override ColorCatalog.ColorIndex EquipmentColorIndex => ColorCatalog.ColorIndex.LunarItem;
+        public override void Hooks()
+        {
+            var target = typeof(EquipmentIcon).GetMethod(nameof(EquipmentIcon.SetDisplayData), System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+            var destination = typeof(ArmletOfMordiggian).GetMethod(nameof(ModifyDisplayData), System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+            new Hook(target, destination, this);
+
+            On.RoR2.CharacterBody.OnEquipmentLost += OnEquipmentLost;
+        }
         public override void Init(ConfigFile config)
         {
             CreateConfig(config);
             CreateLanguageTokens();
             CreateEquipmentDef();
+            Hooks();
         }
 
         public float DamageBonus;
@@ -43,20 +54,101 @@ namespace ROTA2.Equipment
         {
             if (slot && HasThisEquipment(slot.characterBody))
             {
-                EquipmentState state = new()
+                var behavior = slot.characterBody.GetComponent<ArmletOfMordiggianBehavior>();
+                if (behavior)
                 {
-                    equipmentIndex = ArmletOfMordiggianToggled.Instance.EquipmentDef.equipmentIndex,
-                    chargeFinishTime = Run.FixedTimeStamp.now + ArmletCooldown,
-                    charges = 0,
-                    equipmentDef = ArmletOfMordiggianToggled.Instance.EquipmentDef
-                };
-                slot.inventory.SetEquipment(state, slot.activeEquipmentSlot);
+                    slot.characterBody.RemoveBuff(ArmletOfMordiggianBuff.Instance.BuffDef);
+                    slot.characterBody.RemoveBuff(RoR2Content.Buffs.HealingDisabled);
+                    UnityEngine.Object.Destroy(slot.characterBody.GetComponent<ArmletOfMordiggianBehavior>());
+                    Util.PlaySound("ArmletOfMordiggianOff", slot.characterBody.gameObject);
+                }
+                else
+                {
+                    slot.characterBody.AddBuff(ArmletOfMordiggianBuff.Instance.BuffDef);
+                    slot.characterBody.AddBuff(RoR2Content.Buffs.HealingDisabled);
+                    slot.characterBody.gameObject.AddComponent<ArmletOfMordiggianBehavior>();
+                    Util.PlaySound("ArmletOfMordiggianOn", slot.characterBody.gameObject);
+                }
             }
 
             return true;
         }
+
+        private string OffPath = "ROTA2.Icons.armlet_of_mordiggian.png";
+        private string OnPath = "ROTA2.Icons.armlet_of_mordiggian_toggled.png";
+        private void ModifyDisplayData(Action<EquipmentIcon, EquipmentIcon.DisplayData> orig, EquipmentIcon self, EquipmentIcon.DisplayData data)
+        {
+            orig(self, data);
+            if (self && self.targetEquipmentSlot && self.targetEquipmentSlot.characterBody && self.currentDisplayData.equipmentDef == EquipmentDef)
+            {
+                Texture new_texture = null;
+                var behavior = self.targetEquipmentSlot.characterBody.GetComponent<ArmletOfMordiggianBehavior>();
+                if (!behavior)
+                {
+                    new_texture = Plugin.ExtractSprite(OffPath).texture;
+                }
+                else
+                {
+                    new_texture = Plugin.ExtractSprite(OnPath).texture;
+                }
+
+                if (new_texture)
+                {
+                    self.iconImage.texture = new_texture;
+                }
+            }
+        }
+        private void OnEquipmentLost(On.RoR2.CharacterBody.orig_OnEquipmentLost orig, CharacterBody self, EquipmentDef equipmentDef)
+        {
+            if (equipmentDef == EquipmentDef)
+            {
+                self.RemoveBuff(ArmletOfMordiggianBuff.Instance.BuffDef);
+                self.RemoveBuff(RoR2Content.Buffs.HealingDisabled);
+                UnityEngine.Object.Destroy(self.GetComponent<ArmletOfMordiggianBehavior>());
+                Util.PlaySound("ArmletOfMordiggianOff", self.gameObject);
+            }
+
+            orig(self, equipmentDef);
+        }
+
+        public class ArmletOfMordiggianBehavior : MonoBehaviour
+        {
+            HealthComponent health;
+            float elapsed = 0.0f;
+            float tick = 0.2f;
+
+            void Awake()
+            {
+                health = GetComponent<HealthComponent>();
+            }
+
+            void Update()
+            {
+                elapsed += Time.deltaTime;
+                if (elapsed >= tick)
+                {
+                    elapsed -= tick;
+                    DamageTypeCombo combo = new()
+                    {
+                        damageSource = DamageSource.Equipment,
+                        damageType = DamageType.NonLethal | DamageType.BypassBlock | DamageType.Silent
+                    };
+                    DamageInfo info = new()
+                    {
+                        damage = health.fullCombinedHealth * ArmletOfMordiggian.Instance.MaximumHealthLostPerSecond / 100.0f * tick,
+                        procCoefficient = 0.0f,
+                        damageType = combo,
+                        damageColorIndex = DamageColorIndex.Bleed,
+                        position = health.body.transform.position
+                    };
+                    health.TakeDamage(info);
+                }
+            }
+        }
     }
 
+    /* From a more difficult time...
+     * 
     public class ArmletOfMordiggianToggled : EquipmentBase<ArmletOfMordiggianToggled>
     {
         public override string EquipmentName => "Armlet of Mordiggian (Toggled)";
@@ -139,41 +231,6 @@ namespace ROTA2.Equipment
 
             orig(self, equipmentDef);
         }
-
-        public class ArmletOfMordiggianBehavior : MonoBehaviour
-        {
-            HealthComponent health;
-            float elapsed = 0.0f;
-            float tick = 0.2f;
-
-            void Awake()
-            {
-                health = GetComponent<HealthComponent>();
-            }
-
-            void Update()
-            {
-                elapsed += Time.deltaTime;
-                if (elapsed >= tick)
-                {
-                    elapsed -= tick;
-                    DamageTypeCombo combo = new()
-                    {
-                        damageSource = DamageSource.Equipment,
-                        damageType = DamageType.NonLethal | DamageType.BypassBlock | DamageType.Silent
-                    };
-                    DamageInfo info = new()
-                    {
-                        damage = health.fullCombinedHealth * ArmletOfMordiggianToggled.Instance.MaximumHealthLostPerSecond / 100.0f * tick,
-                        procCoefficient = 0.0f,
-                        damageType = combo,
-                        damageColorIndex = DamageColorIndex.Bleed,
-                        inflictor = this.gameObject,
-                        position = health.body.transform.position
-                    };
-                    health.TakeDamage(info);
-                }
-            }
-        }
     }
+    */
 }